Block을 sealed interface로 변경 Block, CleanBlock, MineBlock을 하나의 파일로 합침
