--- src/main/kotlin/minesweeper/CleanBlock.kt
@@ -1,13 +1,0 @@
-package minesweeper
-
-class CleanBlock(
-    private var _nearbyMineCount: Int = 0
-) : Block {
-    override fun getNearbyMineCount(): Int {
-        return _nearbyMineCount
-    }
-
-    override fun setNearbyMineCount(mineCount: Int) {
-        _nearbyMineCount = mineCount
-    }
-}
