--- src/main/kotlin/minesweeper/MineBlock.kt
@@ -1,10 +1,0 @@
-package minesweeper
-
-class MineBlock(
-    private var _nearbyMineCount: Int = 0
-) : Block {
-    override fun setNearbyMineCount(mineCount: Int) {}
-    override fun getNearbyMineCount(): Int {
-        return _nearbyMineCount
-    }
-}
